Adicionando funcionalidade de login e refatorando

+++ src/main/java/com/softwarearchitecture/smartsale/adapter/controllers/DeleteProductController.java
@@ -1,7 +1,7 @@
 package com.softwarearchitecture.smartsale.adapter.controllers;
 
-import com.softwarearchitecture.smartsale.entities.Product;
 import com.softwarearchitecture.smartsale.useCase.DeleteProductUseCaseInterface;
+import com.softwarearchitecture.smartsale.utils.VerifyUser;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,8 +15,11 @@ public class DeleteProductController {
     private DeleteProductUseCaseInterface useCase;
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<Object> deleteProduct(@PathVariable("id") Long id) {
-       useCase.deleteProduct(id);
+    public ResponseEntity<Object> deleteProduct(@PathVariable("id") Long id) throws Exception {
+
+        VerifyUser.checkAdmin();
+
+        useCase.deleteProduct(id);
 
         return ResponseEntity.noContent().build();
     }
